refactor sorting into extra directory
