+++ packages/core/src/index.ts
@@ -2,6 +2,7 @@ export * from './constants';
 export * from './utils';
 export * from './helper';
 export * from './types';
+export * from './sorting';
 
 export * from './components/mediaToggle';
 export * from './components/mediaDeviceSelect';
