+++ packages/core/src/sorting/index.ts
@@ -1,0 +1,1 @@
+export { sortParticipantsByVolume } from './sort-participants';
